Pkg: refactor base/pkg.jl to just cd and calls an entry point.
This avoids all the annoying nonsense of having to avoid recursive
calls to Dir.cd and reasoning about if you are or are not in the
~/.julia directory yet.
