+++ base/sysimg.jl
@@ -209,8 +209,6 @@ include("deprecated.jl")
 # git utils & package manager
 include("git.jl")
 include("pkg.jl")
-include("pkg1.jl")
-const Pkg2 = Pkg
 
 # base graphics API
 include("graphics.jl")
